geo_tool/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
openai>=1.50.0
requests>=2.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
hypothesis>=6.100.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
